% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{csmfSummaryCSV}
\alias{csmfSummaryCSV}
\title{Format InterVA5 & InSilicoVA summary objects to write as CSV file}
\usage{
csmfSummaryCSV(x)
}
\arguments{
\item{x}{A summary of object returned from openVA::codeVA()}
}
\description{
Format InterVA5 & InSilicoVA summary objects to write as CSV file
}
