% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{indivCOD}
\alias{indivCOD}
\title{Get the CODs assigned to each individual in the data}
\usage{
indivCOD(x, top = 3)
}
\arguments{
\item{x}{An fitted object returned from openVA::codeVA()}

\item{top}{An integer specifying the number of top causes included
in the results.}
}
\value{
A matrix with the death ID, top causes, and associated
probability for the top causes.
}
\description{
Get the CODs assigned to each individual in the data
}
