% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch openVA shiny app in internet browser}
\usage{
launchApp()
}
\description{
Launch openVA shiny app in internet browser
}
\examples{
\dontrun{
library(openVAapp)
launchApp()
}
}
