% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sepVALog}
\alias{sepVALog}
\title{Separate InterVA5 log messages by age and sex.}
\usage{
sepVALog(x, names_runs, log_file)
}
\arguments{
\item{x}{A fitted object returned from openVA::codeVA()}

\item{names_runs}{Character vector with names of demographic
groups to include in the stratification of results.
Possible values include: all, male, female, neonate, child,
adult, mNeonate, mChild, mAdult, fNeonate, fChild, and fAdult.}

\item{log_file}{String with path to log file.}
}
\description{
Separate InterVA5 log messages by age and sex.
}
