% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sepVAResults}
\alias{sepVAResults}
\title{Separate InterVA5 results by age and sex.}
\usage{
sepVAResults(x)
}
\arguments{
\item{x}{A fitted object returned from openVA::codeVA()}
}
\description{
Separate InterVA5 results by age and sex.
}
